requirements.txt
aio-pika==6.7.1
aiocron==1.3
aiohttp==3.7.4.post0
aioredis==1.3.1
aiormq==3.3.1
appdirs==1.4.4
aspy.refactor-imports==2.1.1
async-timeout==3.0.1
asyncpg==0.21.0
attrs==20.3.0
autoflake==1.4
bandit==1.7.0
black==20.8b1
cached-property==1.5.2
cfgv==3.2.0
chardet==4.0.0
click==7.1.2
croniter==1.0.6
distlib==0.3.1
eradicate==2.0.0
filelock==3.0.12
flake8==3.8.4
flake8-annotations==2.4.1
flake8-annotations-coverage==0.0.4
flake8-bandit==2.1.2
flake8-broken-line==0.3.0
flake8-bugbear==20.11.1
flake8-builtins==1.5.3
flake8-commas==2.0.0
flake8-comprehensions==3.3.1
flake8-debugger==4.0.0
flake8-eradicate==1.0.0
flake8-executable==2.1.1
flake8-fixme==1.1.1
flake8-future-import==0.4.6
flake8-mutable==1.2.0
flake8-plugin-utils==1.3.1
flake8-polyfill==1.0.2
flake8-pyi==20.10.0
flake8-pytest==1.3
flake8-pytest-style==1.3.0
flake8-string-format==0.3.0
flake8-todo==0.7
flake8-unused-arguments==0.0.6
future==0.18.2
gitdb==4.0.5
GitPython==3.1.11
h11==0.11.0
hiredis==1.1.0
identify==1.5.10
idna==2.10
iniconfig==1.1.1
mccabe==0.6.1
multidict==5.1.0
mypy-extensions==0.4.3
natsort==7.1.1
nodeenv==1.5.0
packaging==20.8
pamqp==2.3.0
pathspec==0.8.1
pbr==5.5.1
pluggy==0.13.1
pre-commit==2.9.3
py==1.10.0
pycodestyle==2.6.0
pydantic==1.7.3
pyflakes==2.2.0
pyparsing==2.4.7
pytest==6.2.1
pytest-asyncio==0.14.0
pytest-django==4.1.0
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2021.1
pyupgrade==2.7.4
PyYAML==5.3.1
regex==2020.11.13
reorder-python-imports==2.3.6
simple-print==0.5
six==1.15.0
smmap==3.0.4
starlette==0.14.1
stevedore==3.3.0
termcolor==1.1.0
tokenize-rt==4.0.0
toml==0.10.2
typed-ast==1.4.1
typing-extensions==3.7.4.3
tzlocal==2.1
ujson==4.0.1
uvicorn==0.13.2
uvloop==0.14.0
virtualenv==20.2.2
watchgod==0.6
yarl==1.6.3
yesqa==1.2.2
